🐛 fix(table): No more transparent formitem
