+++ packages/form/src/components/Checkbox/index.tsx
@@ -3,7 +3,7 @@ import { Checkbox } from 'antd';
 import ProField from '@ant-design/pro-field';
 import { ProSchema } from '@ant-design/pro-utils';
 import { CheckboxGroupProps, CheckboxProps } from 'antd/lib/checkbox';
-import { createField } from '../../BaseForm';
+import createField from '../../BaseForm/createField';
 import { ProFormItemProps } from '../../interface';
 
 export type ProFormCheckboxGroupProps = ProFormItemProps<CheckboxGroupProps> & {
