+++ packages/form/src/components/DateRangePicker/index.tsx
@@ -2,8 +2,7 @@ import React from 'react';
 import ProField from '@ant-design/pro-field';
 import { DatePickerProps } from 'antd/lib/date-picker';
 import { ProFormItemProps } from '../../interface';
-import { createField } from '../../BaseForm';
-
+import createField from '../../BaseForm/createField';
 const valueType = 'dateRange';
 
 /**
