+++ packages/form/src/components/Select/index.tsx
@@ -3,8 +3,7 @@ import { SelectProps } from 'antd/lib/select';
 import ProField from '@ant-design/pro-field';
 import { ProSchema } from '@ant-design/pro-utils';
 import { ProFormItemProps } from '../../interface';
-import { createField } from '../../BaseForm';
-
+import createField from '../../BaseForm/createField';
 type ProFormSelectProps = ProFormItemProps<SelectProps<any>> & {
   valueEnum?: ProSchema['valueEnum'];
   request?: ProSchema['request'];
