+++ packages/form/src/components/Text/index.tsx
@@ -2,8 +2,7 @@ import React from 'react';
 import { InputProps, PasswordProps } from 'antd/lib/input';
 import ProField from '@ant-design/pro-field';
 import { ProFormItemProps } from '../../interface';
-import { createField } from '../../BaseForm';
-
+import createField from '../../BaseForm/createField';
 const valueType = 'text';
 /**
  * 文本组件
