+++ packages/form/src/components/TimePicker/index.tsx
@@ -3,8 +3,7 @@ import React from 'react';
 import ProField from '@ant-design/pro-field';
 import { DatePickerProps } from 'antd/lib/date-picker';
 import { ProFormItemProps } from '../../interface';
-import { createField } from '../../BaseForm';
-
+import createField from '../../BaseForm/createField';
 const valueType = 'time';
 /**
  * 时间选择组件
