+++ packages/table/src/Form/index.tsx
@@ -249,7 +249,7 @@ export const proFormItemRender: (props: {
     label: getTitle(),
     request,
     params,
-    ...formItemProps,
+    formItemProps,
   });
   if (!dom) {
     return null;
